backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.2
python-dateutil==2.8.2
python-multipart==0.0.6
bcrypt==4.0.1
python-jose[cryptography]==3.3.0 